change airtime data type to unsigned int

+++ src/packet_analyzer.h
@@ -24,7 +24,7 @@ struct MCS_radiotap_header {
 
 struct arguments{
 	pcap_dumper_t *dumper;
-	float airtime;
+	unsigned int airtime;
 	int duration; //duration to capture
 	time_t start; //time to start capturing
 	pcap_t *handler;
